implementación de middlelware y validación de roles en departamentos

+++ src/pages/api/departamentos/[id].js
@@ -1,7 +1,7 @@
 import db from '@/database/config';
-export default function handler(req, res) {
+import authMiddlelware from '@/helpers/middleware';
+function handler(req, res) {
   const { id } = req.query; // Obtener el ID del departamento de la consulta
-
 
   if (req.method === 'GET') {
     db.query('SELECT * FROM Departamento WHERE idDepartamento = ?', [id], (error, results, fields) => {
@@ -16,6 +16,10 @@ export default function handler(req, res) {
       }
     });
   } else if (req.method === 'PUT') {
+    // Verificar si el usuario tiene permisos para actualizar el departamento
+    if (req.user.idRolUsuario !== 1) {
+      return res.status(401).json({ error: 'No tiene permisos para actualizar el departamento' });
+    }
     const { nombreDepartamento, descripcion, limiteLineas } = req.body;
     db.query('UPDATE Departamento SET nombreDepartamento = ?, descripcion = ?, limiteLineas = ? WHERE idDepartamento = ?', [nombreDepartamento, descripcion, limiteLineas, id], (error, results, fields) => {
       if (error) {
@@ -25,6 +29,10 @@ export default function handler(req, res) {
       }
     });
   } else if (req.method === 'DELETE') {
+    // Verificar si el usuario tiene permisos para eliminar el departamento
+    if (req.user.idRolUsuario !== 1) {
+      return res.status(401).json({ error: 'No tiene permisos para eliminar el departamento' });
+    }
     db.query('DELETE FROM Departamento WHERE idDepartamento = ?', [id], (error, results, fields) => {
       if (error) {
         res.status(500).json({ error: 'Error al eliminar el departamento' });
@@ -36,3 +44,5 @@ export default function handler(req, res) {
     res.status(404).json({ error: 'Ruta no encontrada' });
   }
 }
+
+export default authMiddlelware(handler);
